# bubuka_pro/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для микросервиса",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rels_dls/{rel_dl_id}/": {
            "get": {
                "tags": [
                    "Releases_Downloads"
                ],
                "summary": "Получение записи",
                "operationId": "068d7a7e62f8a4aeadf85e159ddbd4f2",
                "parameters": [
                    {
                        "name": "rel_dl_id",
                        "in": "path",
                        "description": "Id записи скаченного релиза",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Release_Download"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Releases_Downloads"
                ],
                "summary": "Удаление записи",
                "operationId": "af28dc3ae91edc4332e60e5a83097a58",
                "parameters": [
                    {
                        "name": "rel_dl_id",
                        "in": "path",
                        "description": "Id записи скаченного релиза",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/404"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/500"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/rels_dls/": {
            "get": {
                "tags": [
                    "Releases_Downloads"
                ],
                "summary": "Получение записей о скачавших пользователях",
                "operationId": "49d7b0753feee226c2b4a8fc4cded648",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Кол-во записей на странице",
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Кол-во пропусков",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Release_Download"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Статус кода",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "description": "Сообщение ошибки",
                                            "properties": {
                                                "limit": {
                                                    "description": "Возможные ошибки валидации поля limit",
                                                    "type": "string",
                                                    "example": [
                                                        "The limit field is required when offset are present.",
                                                        "The limit field must be a integer"
                                                    ]
                                                },
                                                "offset": {
                                                    "description": "Возможные ошибки валидации поля offset",
                                                    "type": "string",
                                                    "example": "The offset field must be a integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "code": {
                                            "description": "Код ответа",
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Release_Download": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "body": {
                        "properties": {
                            "releases_downloads/release_download": {
                                "properties": {
                                    "id": {
                                        "description": "Id записи",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "release_id": {
                                        "description": "Id релиза",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "ip": {
                                        "description": "Ip адрес",
                                        "type": "string",
                                        "example": "127.0.0.1"
                                    },
                                    "user_agent": {
                                        "description": "User-agent",
                                        "type": "string",
                                        "example": "Mozilla/5.0(X11; Ubuntu; Linux x86_64; rv:122.0) Gecko/20100101 Firefox/122.0"
                                    },
                                    "utm": {
                                        "description": "Спеу метки",
                                        "type": "string",
                                        "example": "{\\'Source\\':null,\\'Medium\\':null,\\'Campaign\\':null}"
                                    },
                                    "created_at": {
                                        "description": "Дата создания",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-12-27T09:59:49.000000Z"
                                    },
                                    "updated_at": {
                                        "description": "Дата изменения",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-12-27T09:59:49.000000Z"
                                    },
                                    "deleted_at": {
                                        "description": "Дата удаления",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "404": {
                "properties": {
                    "status": {
                        "description": "Статус ответа",
                        "type": "boolean",
                        "example": "false"
                    },
                    "message": {
                        "description": "Сообщение ответа",
                        "type": "string",
                        "example": "Запись не найдена"
                    },
                    "code": {
                        "description": "Код ответа",
                        "type": "integer",
                        "example": "404"
                    }
                },
                "type": "object"
            },
            "500": {
                "properties": {
                    "status": {
                        "description": "Статус ответа",
                        "type": "boolean",
                        "example": "false"
                    },
                    "message": {
                        "description": "Сообщение ответа",
                        "type": "string",
                        "example": "Ошибка сервера"
                    },
                    "code": {
                        "description": "Код ответа",
                        "type": "integer",
                        "example": "500"
                    }
                },
                "type": "object"
            },
            "401": {
                "properties": {
                    "status": {
                        "description": "Статус ответа",
                        "type": "boolean",
                        "example": "false"
                    },
                    "message": {
                        "description": "Сообщение ответа",
                        "type": "string",
                        "example": [
                            "Token not provided",
                            "The token has been blacklisted"
                        ]
                    },
                    "code": {
                        "description": "Код ответа",
                        "type": "integer",
                        "example": "401"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}